+++ Lib/fontbakery/profiles/googlefonts.py
@@ -1861,14 +1861,16 @@ def com_google_fonts_check_metadata_has_regular(family_metadata):
 
     if has_regular_style(family_metadata):
         yield PASS, "Family has a Regular style."
-    elif family_metadata.name.endswith(" One") and not has_regular_style(family_metadata):
-        yield PASS, "Family is single-weight non-regular ('One' appended to family name)."
     else:
         yield FAIL,\
               Message("lacks-regular",
                       "This family lacks a Regular"
                       " (style: normal and weight: 400)"
-                      " as required by Google Fonts standards.")
+                      " as required by Google Fonts standards."
+                      " "
+                      " If family consists of a single-weight non-Regular style only,"
+                      " consider the Google Fonts specs for this case:"
+                      " https://github.com/googlefonts/gf-docs/tree/main/Spec#single-weight-families")
 
 
 @check(